# docs/openapi.yml
openapi: 3.1.0
info:
  title: fast-auth
  description: A simple, extensible authentication library for Axum with JWT and refresh tokens
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  version: 0.1.0
paths:
...
  /auth/organizations:
    get:
      tags:
      - auth
      operationId: organizations_list
      responses: {}
      security:
      - sessionCookie: []
      - bearerApiKey: []
    post:
      tags:
      - auth
      operationId: organizations_create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationCreateRequest'
        required: true
      responses: {}
      security:
      - sessionCookie: []
      - bearerApiKey: []
  /auth/organizations/current:
    post:
      tags:
      - auth
      operationId: organizations_switch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationSwitchRequest'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
      security:
      - sessionCookie: []
      - bearerApiKey: []
  /auth/organizations/invites/accept:
    post:
      tags:
      - auth
      operationId: organization_invite_accept
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationInviteAcceptRequest'
        required: true
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
        '500':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthErrorResponse'
      security:
      - sessionCookie: []
      - bearerApiKey: []
  /auth/organizations/{organization_id}:
    get:
      tags:
      - auth
      operationId: organizations_get
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses: {}
      security:
      - sessionCookie: []
      - bearerApiKey: []
    delete:
      tags:
      - auth
      operationId: organizations_delete
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses: {}
      security:
      - sessionCookie: []
      - bearerApiKey: []
    patch:
      tags:
      - auth
      operationId: organizations_update
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationUpdateRequest'
        required: true
      responses: {}
      security:
      - sessionCookie: []
      - bearerApiKey: []
  /auth/organizations/{organization_id}/invites:
    get:
      tags:
      - auth
      operationId: organization_invites_list
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses: {}
      security:
      - sessionCookie: []
      - bearerApiKey: []
    post:
      tags:
      - auth
      operationId: organization_invite_create
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationInviteCreateRequest'
        required: true
      responses: {}
      security:
      - sessionCookie: []
      - bearerApiKey: []
  /auth/organizations/{organization_id}/invites/{invite_id}:
    delete:
      tags:
      - auth
      operationId: organization_invite_revoke
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: invite_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses: {}
      security:
      - sessionCookie: []
      - bearerApiKey: []
  /auth/organizations/{organization_id}/members:
    get:
      tags:
      - auth
      operationId: organization_members_list
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses: {}
      security:
      - sessionCookie: []
      - bearerApiKey: []
  /auth/organizations/{organization_id}/members/{member_user_id}:
    delete:
      tags:
      - auth
      operationId: organization_member_delete
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: member_user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses: {}
      security:
      - sessionCookie: []
      - bearerApiKey: []
    patch:
      tags:
      - auth
      operationId: organization_member_update
      parameters:
      - name: organization_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: member_user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrganizationRoleUpdateRequest'
        required: true
      responses: {}
      security:
      - sessionCookie: []
      - bearerApiKey: []
...
components:
  schemas:
...
    AuthErrorResponse:
      type: object
      description: Standard error response body returned by auth endpoints.
      required:
      - error
      properties:
        error:
          type: string
          description: Human-readable error message.
    AuthResponse:
      type: object
      description: Auth response body. Tokens are set as httpOnly cookies.
      required:
      - user
      - organization
      - auth_role
      properties:
        auth_role:
          type: string
        organization:
          $ref: '#/components/schemas/OrganizationResponse'
        user:
          $ref: '#/components/schemas/UserResponse'
...
    OrganizationCreateRequest:
      type: object
      description: Organization create request.
      required:
      - name
      properties:
        name:
          type: string
          description: Organization name.
...
    OrganizationInviteAcceptRequest:
      type: object
      description: Invitation accept request.
      required:
      - token
      properties:
        token:
          type: string
          description: Plaintext invitation token.
    OrganizationInviteCreateRequest:
      type: object
      description: Invitation create request.
      required:
      - email
      - role
      properties:
        email:
          type: string
          description: Invited email.
        role:
          $ref: '#/components/schemas/OrganizationRole'
          description: Invited role.
...
    OrganizationResponse:
      type: object
      description: Active organization returned in auth responses.
      required:
      - id
      - name
      - role
      properties:
        id:
          type: string
        name:
          type: string
        role:
          $ref: '#/components/schemas/OrganizationRole'
    OrganizationRole:
      type: string
      description: |-
        Organization membership role exposed by auth.

        This role is embedded in [`CurrentUser`], organization membership responses,
        and invitations so backends and HTTP handlers can share one stable
        authorization vocabulary.
      enum:
      - owner
      - admin
      - member
    OrganizationRoleUpdateRequest:
      type: object
      description: Membership role update request.
      required:
      - role
      properties:
        role:
          $ref: '#/components/schemas/OrganizationRole'
          description: New organization role.
    OrganizationSwitchRequest:
      type: object
      description: Active organization switch request.
      required:
      - organization_id
      properties:
        organization_id:
          type: string
          format: uuid
          description: Next active organization id.
    OrganizationUpdateRequest:
      type: object
      description: Organization update request.
      required:
      - name
      properties:
        name:
          type: string
          description: Updated organization name.
...
    UserResponse:
      type: object
      description: User data returned in auth responses.
      required:
      - id
      - email
      properties:
        email:
          type: string
        email_confirmed_at:
          type:
          - string
          - 'null'
        id:
          type: string
  securitySchemes:
    bearerApiKey:
      type: apiKey
      in: header
      name: Authorization
    sessionCookie:
      type: apiKey
      in: cookie
      name: access_token
...